Improve cforg/cfspace finalizers
We should only remove their finalizers if the undelrying namespace is
gone (i.e. delete returns a NotFound error). This way we are preventing
the case where the cforg/cfspace is gone, but the namespace is stuck in
Terminating, because some object could not get finalized.

The tests for cforgs and cfspace finalization have been moved to the
crds test suite as it runs on kind. It turns out that when using envtest
namespaces never go away, but stay Terminating forever.

https://book.kubebuilder.io/reference/envtest.html#namespace-usage-limitation

Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>