Do not require the CFDomain to be present to finalize CFRoute
This prevents a problem where CFRoutes are stuck in the finalization
state because their CFDomain has been deleted. Note, since CFDomain
finalization delete its associated CFRoutes, we can't see how this could
happen, but it did on a CI cluster, so we're addressing it.

Instead of requiring the CFDomain to be present in order to finalize the
CFRoute, use the FQDN stored in the context to locate the HTTPProxy
resources.

If FQDN is empty, the reconciler probably hasn't modified any HTTPProxy
resources which would need cleaning up.

Issue: https://github.com/cloudfoundry/korifi/issues/2591
Co-authored-by: Giuseppe Capizzi <gcapizzi@vmware.com>