Revert movement of BuildWorkload and Build finalizers to webhook
This partially reverts commit fbbbd46e403dc91df0b9d7e1322df00a3b26e91e.

The reverted changes make an incorrect assumption: that only one builder
will be running on a cluster at a time. There are already real world
deployments of korifi (i.e. tas-adapter) that have kpack-image-builder
running concurrently with other builders. The reverted code only works
if kpack-image-builder is guaranteed to be the only builder running on a
cluster.

kpack-image-builder may be able to use a webhook to handle finalizers,
but only if it can establish ownership of the resources before adding a
finalizer. Otherwise, we end up in a situation where other builders have
to remove the kpack-image-builder finalizer from BuildWorkloads and
Builds that they own