kpack-image-builder must only add finalizers to its BuildWorkloads
Add a new `SetPolicy` for the BuildWorkload finalizer in
kpack-image-builder which checks the BuildName on the spec, just like
its reconciler does. This way multiple builders can co-exist in korifi
without interfering.

This requires the signature of SetPolicy to be changed to give access to
the Spec of the object. We now use an unstructured object which can be
easily converted to an actual resource type if required.