Retry acquiring lease lock when NotFound encountered
This addresses a flake where a resource is created and quickly renamed
in our integration tests. We suspect the 2 etcd instances are not yet
sync'ed when we attempt to modified the lease created with the original
resource name. We get a NotFound error. By allowing a limited number of
retries with an exponential backoff, we should deal with this condition.

Issue: https://github.com/cloudfoundry/korifi/issues/2587