Fix flake: wait for a task being initialized
When altering the status conditions, if the condititions array is empty,
it is going to be replaced with a new instance. As we have both the
reconciler and the test running in parallel, setting the succeeded
condition from the test might be overwritten by the reconciler by
replacing conditions with an array containing the initialised condition.

By awaiting for the initialised condition in the test we ensure that the
race is avoided.

Co-authored-by: Danail Branekov <danailster@gmail.com>
Co-authored-by: Kieron Browne <kbrowne@vmware.com>