Store namespace and name of owner resource when reserving a name
This annotates name leases with the name and namespace of the resource the
name is being used for. This will help us address some issues raising
from webhooks being invoked more than once.

This also introduces the `UniqueClientObject` interface which we
implemented for every resource that has a unique name validated via the
`NameRegistry`. This simplifies the signatures for the
`DuplicateValidator` and allows each type to establish their own naming
scheme and validation error messages.

Issue: https://github.com/cloudfoundry/korifi/issues/2628
Co-authored-by: Giuseppe Capizzi <gcapizzi@vmware.com>