Tolerate repeated validation for the same name change
It is possible that the k8s API does not complete the name change
request after calling the validation webhook for it (see
https://kubernetes.io/docs/reference/access-authn-authz/extensible-admission-controllers/#side-effects).
Since our name change validation mutates `Lease` resources, it is
side-effecting, and we need to be aware of when the leases might get out
of sync with the resources that own them.

This change addresses the particular problem we see when the API retries
a rename patch. On the first attempt, the validation removes the lease
on the original name and creates a lease for the new name. The the API
doesn't persist the name change. It then tries the again. This time the
validation fails as the lease for the old name can't be found.

We now work around this particular case. If the old lease cannot be
found, but a lease for the new name exists and is owned by the resource
in question, we allow the update request.

We don't appear to have problems with the create and delete validation.

Issue: https://github.com/cloudfoundry/korifi/issues/2628
Co-authored-by: Giuseppe Capizzi <gcapizzi@vmware.com>