Build workloads do not own custom kpack builders via ControllerReference
Custom kpack builders (the ones that are created on the fly for build
workloads with buildpacks specified) are shared by build workloads
that have the same set of buildpacks. Therefore, setting
ControllerReference from the build workload to the builder is incorrect
as the ControllerReference can only be set once.

Issue: https://github.com/cloudfoundry/korifi/issues/2645