Allow the name registry to get, list and watch leases
In order to check whether a name has been already reserved by the object
being updated (i.e. the lease for the old name does not exist, but a
lease owned by the updated object for the new name does), the name
registry has to be allowed to get leases.

However, when the name registry is used in the context of
webhooks/controllers, once the lease is got, the webhook's client cache
starts watching and listing the leases. Therefore, the webhook user has
to be permitted to watch and list leases as well.

This requirement was not covered by our integration tests as webhooks
used to be run with the admin client (that is configured by env test).
In order to close this gap, webhooks are now run as an user that is bound
to the controllers role.

Co-authored-by: Danail Branekov <danailster@gmail.com>