Make the jobs handler generic over the job type
The handler now receives a map from job type to a `DeletionRepository`,
which is all it needs to implement the deletion job logic. Supporting
deletion jobs for new resources is now only a matter of implementing the
`DeletionRepository` interface for the resource repository, and then add
an entry to the map.