Fix GetDeletedAt for orgs
The GetDeletedAt method now delegates to GetOrg. This is important since
orgs are special from rbac point of view since in order to know if we
have permission in an org we have to have role bindings to some role in
that org. This is why GetOrg is implemented via List-ing orgs and
filtering by guid. Not following that pattern results in any user seeing
any orgs.

Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>