Fix deploy-on-kind with docker hub as registry
When using deploy-on-kind we default to using the local registry. In
case we want a custom registry we should export the following vars:

DOCKER_SERVER
DOCKER_USERNAME
DOCKER_PASSWORD
REPOSOTORY_PREFIX
KPACK_BUILDER_REPOSITORY

In the INSTALL.kind.md we advise that people use dockerhub as the
registry, but setting DOCKER_SERVER to index.docker.io would result in a
wrong image-registry-credentials secret as the .dockerconfigjson wants
https://index.docker.io/v1/ as the host name. On the other hand
REPOSOTORY_PREFIX should be `index.dpcker.io/username/ and
KPACK_BUILDER_REPOSITORY should be
`index.docker.io/username/kpack-builder`, i.e. they do not have the
https scheme and the /v1 path suffix. The docker hub config json is very
picky about the server name and we need special treatment for this use
case to work

Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>