Use uncached client in the name registry integration test
The name registry is an utility that is used in the context of
controllers which in case of outdated cache just retry reconciliation.

However, the integration test is not such a context, therefore using a
caching client does not make sense. Furthermore, doing so could cause
flakes due to a cache miss, for example https://ci.korifi.cf-app.com/teams/main/pipelines/main/jobs/run-tests-main/builds/950