Use the admin client for e2e tests
Now that we are not testing permissions with e2e tests, there is no
point to have both admin and unprivileged clients, we could just use the
admin one. That reduces setup complexity significantly.

Also, there is no need to assign roles to the unprivileged user anymore
which reduces `BeforeEach`es a lot.

In the cf_user_test we create a service account on the fly to test that
users without any role bindings are getting a warning when calling the
API. This paves the road for future improvements where tests create the
users they need themselves instead of relying on external setup

Co-authored-by: Danail Branekov <danailster@gmail.com>
Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>