Sync client now supports synchronous deletes
Depending on the object type `Delete` would delete the object and then
a) if the object is namespace - await its `DeletionTimestamp` to get set
- test env does not actually deletes namespaces
b) if the object is CFOrg or CFSpace with existing underlying namespace
- await their `DeletionTimestamps` to be set as they cannot be deleted
  while the namespace is around (due to testenv)
c) otherwise - await `Get` to start returning a not
found error

There are few cases when synchronous deletion is not desired though, for
example, a test verifying that a webhook would allow deletion. The
object webhook is installed by a defaulter, but in the webhook test
suite there is no object controller to eventually remove the finalizer,
thus the object remains. In those cases a standard client is used.

Co-authored-by: Danail Branekov <danailster@gmail.com>