Fix expired CFTask deletion test flake
Instead of messing up with the CFTask status that is supposed to be a
repsponsibility of the CFTask controller itself, set the status on the
task workload. We then expect the CFTask controller to notice the workload
change and self-rechedule itself to delete the CFTask in a while.

fixes https://github.com/cloudfoundry/korifi/issues/2873

Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>
Co-authored-by: Danail Branekov <danailster@gmail.com>