Create BTP service instance/bindings when reconciling a managed service
In BTP services terms, the service instance registers the service with
BTP and the BTP binding creates the secret to be used by apps.
Therefore, when reconciling managed services that want to consume BTP
services, we need to create both the BTP service instance and the BTP
service binding and let the BTP controller populate the secret
referenced by the CFServiceInstance.
Thus we could use the standard CFServiceBinding controller
implementation that mounts the secret into the app container (via
servicebinding.io artifacts)

Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>