Fix a CI flake in cfBuild controller tests
Link to failed build: https://ci.korifi.cf-app.com/teams/main/pipelines/main/jobs/run-tests-periodic/builds/12261#L6500636d:59

This flake has occurred a couple of times during the weekend. It was
most probably caused by the cfbuild controller seeing the build success
status too soon, causing it to not delegate to the delegate reconciler
even once. Changing the test to wait for the first invocation of the
delegate reconciler before patching the build status to successfuls
state would prevent this problem. Then it would be safe to assert that
the delegate call count remains stable.

Also  renamed the fakeStatusCondition to delegateInvokedCondition for
clarity