Consider build workload not being ready if ready!=true
This test awaits for the build workload reconciler to notice that the
builder is not ready:

* if the builder status is unknown, the build workload status is also
  set to unknown
* if the builder is not ready, the workload redy status is set to false

In either case, test should assert that the build worklod consistently
does not become ready (i.e. its ready status is not true)
before making the builder ready.

Also, when the builder is not ready, the workload reconciler would
requeue the event in a second in order to avoid the exponential backoff
delay.

Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>