Use `builderReadyCondition.IsTrue()` to check readiness
Previously we used `builderReadyCondition.IsFalse()` to check whether
the builder is ready. However, `IsFalse` returns `false` when the
condition is `nil` which is not what we want - when the ready condition
is `nil` we should consider that the builder is not ready.

Instead, we now use `builderReadyCondition.IsTrue()` which only returns
`true` if the ready condition is set, and its status is `True`.

We also change the way we check for builder ready timeout - we base the
check on the build workload creation time instead of the builder ready
status condition transition time as the condition itself might be nil.

Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>