Ignore org deletion exit code in AfterSuite
Smoke tests have been failing frequently because org deletion times out.
Previous debug output showed that the timeout is caused by org namespace
being not deleted because of app pods do not disappear and remaining
pods are in phase `Pending` (e.g. see
https://ci.korifi.cf-app.com/teams/main/pipelines/main/jobs/deploy-korifi-acceptance/builds/802)

We have found a [k8s bug
report](https://github.com/kubernetes/kubernetes/issues/121435) that
pending pods cannot be deleted. We suspect that in those cases the
cluster is just being slow.

In order to workaround that bug, we simply ignore the org deletion exit
code as org deletion is not really in the focus of the smoke tests. Org
deletion is covered in details by e2e tests.

Also, delete the related debug logs as they are not needed anymore.

fixes #3061

Co-authored-by: Danail Branekov <danailster@gmail.com>
Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>