This branch is an attempt to reproduce a weird problem that we observed with servicebinding.io bindings
The idea is to create a cfservicebinding and a servicebinding.io binding
that uses the cfservicebinding as a service and a simple deployment as a
workolad

Initially we leave the cfservicebinding duck type empty, then create the
servicebinding.io binding, then set the duck type.

In this case the expectation would be that the servicebinding.io becomes
ready eventyally, but we suspect that this would not be true unless we
"touch" the servicebinding.io resource by setting a label on it thus
forcing re-reconciliation.

If the above is confirmed we should report a bug to the
servicebinding.io project

The reproduction is in the form of an e2e test, since it involves
setting a status on a CR, which cannot be done via kubectl

Co-authored-by: Danail Branekov <danailster@gmail.com>
Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>