Switch to dynamic contour provisioning on kind
As of contour 1.29.0,

```
Contour can no longer be configured with a GatewayClass controller name (gateway.controllerName in the config file or ContourConfiguration CRD), as the config field has been removed. Instead, either use a specific Gateway reference (gateway.gatewayRef), or use the Gateway provisioner.
```

This makes static contour provisioning awkward as the contour config has
to be changed and then the contour deployment has to be restarted (see
https://projectcontour.io/docs/1.29/guides/gateway-api/#option-1-statically-provisioned)

We have been considering dynamic provisioning for a while, and maybe it
is time.

In order to implement the switch:
* `install-dependencies.sh` installs the dynamic contour provisioner
* `deploy-on-kind.sh`
    * Configures the `GatewayClass` with a `ContourDeployment` config object
    * The `ContourDeployment` config object is configured to use
      NodePortService envoy network publishing so that instead of a
      `LoadBalancer` service, contour genrates a `NodePort` service.
    * The `NodePort` service generated by contour uses the `Gateway`
      listeners' ports for the `NodePort` (see
      https://github.com/projectcontour/contour/issues/4499) - this cannot
      be changed as of today. The range of `NodePort` ports is 30000-327xx,
      therefore gateway ports are patched from `80` to `32080` and from `443` to
      `32443`. Hopefully this should not affect `LoadBalancer` services