Wait for app to become staged after droplet is set
This is an attempt to fix test flakes like this one:

https://ci.korifi.cf-app.com/teams/main/pipelines/main/jobs/run-e2es-eks-periodic/builds/8092#L66377a10:1977

In this test we are pusing an app and then immediately trygin to create
a task for the pushed app. The problem is that the app may take some
time to become properly STAGED and STARTED and the test is not waiting
for the app

Having said that it is not trivial to wait for the app to become
STARTED, because of the nature of the cf v3 api: According to the
[docs](https://v3-apidocs.cloudfoundry.org/version/3.159.0/index.html#the-app-object)
the `state` field in the app response is representing the app desired
state (not the actual state). Luckily there is an endpoint for getting
the current app droplet, namely

``` GET /v3/apps/{appGUID}/droplets/current ```

By waiting for the droplet to appear we are making sure that no test
would proceed before the app is at least STAGED

Analyzing the logs of the failed test it looks like the test is too
quick to attempt to start a task for the app, before the app has its
droplet set. Therefore waiting for the app to be STAGED should fix this
flake.