Ensure app has reached its desired state
The app state is not a final one. An app can be stopped, then started
and then stopped again. Therefore, relying on the `Ready` condition
solely is not realiable enough.
Setting the desired state on the app spec does not clear the `Ready`
status condition, and thus the `Ready` state might have been set by an
unrelated change.

This change makes the app repository not only wait for the app `Ready`
condition, but also ensure that the app desired and actual state do
match.