An attempt to fix the following flake:
https://ci.korifi.cf-app.com/teams/main/pipelines/main/jobs/run-tests-periodic/builds/17471

The test above invalidates the servicebinding.io binding by setting an
outdated generation on the servicebinding.io object. Then it asserts
that the CFServiceBinding object should eventyally have Ready=false and
consistently stay this way. In the flaky run the `Eventually` clause was
sucessful, but the `Consistently` clause was not. This could potentially
be because the `It` comes in too early before the CFServiceBinding
object becomes `Ready` for the first time. Adding a `Ready` check in the
`BeforeEach` before breaing the servicebinding.io binding should exclude
this possibility

Co-authored-by: Danail Branekov <danailster@gmail.com>