Use `go-functional` in repositories `list` methods
This eliminates quite some boilerplate code around filtering and mapping
k8s resources. `go-functional` also replaces our inhouse filtering
implementation.

Also, this change removes code duplications in methods such as
`GetSomethingByAandB` that list and filter resources. Instead of
implementing the list operation themselves, they delegate to the
relevant `list` method by constructing a list messages that matches the
criteria they need. As a further improvement we could get rid of those
methods and just invoke `repo.List` from the handlers.
As a consequence of this change those methods now return a `NotFound`
error rather than `Unauthorised` when the user is not allowed to list
the resources. However, that is fine as the related api handlers mask
fobidden errors as not found, so no change from API users' point of view

Co-authored-by: Danail Branekov <danailster@gmail.com>
Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>