Generalise including resources' fields
`includes` and `fields` are quite similar concepts. The difference is
that `include` inludes the complete resource, while `fields` select the
fields of the resource to be included.

With this commit we abstract that difference away. `IncludedResource`
now has a `SelectJSONFields` method that allows picking the fields one
want to see in the response. Under the hood `SelectJSONFields` uses json
marshal/unmarshal, tehrefore fields selectors work on the resource json
representation.