Use `EnsurePatch` in OSBAPI client factory tests
The client factory is an utility and its tests are not executed in the
context of the controllers retry loop. Therefore, in order to avoid
flakes from k8s eventual consistency, we use `EnsurePatch` in tests
instead of `k8sClient.Patch` to ensure that the factory sees the state
of the objects the test setup creates.

Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>