Patching reconciler ensures non-empty status reason
Status conditions must have non-empty reason, otherwise the patch
operation is rejected by kubernetes.

When constructing the `NotReadyError`, a reconciler might set the cause
but not be able to determine the reason. Instead of trying to adhere to
always setting the reason in the controllers, the patching reconciler
now ensures that it is defaulted to `Unknown`.

While being here, move the `NotReadyError` type into the `reconcile.go`
file - it logically makes sense to be there