Generate resource guids in webhook
This way korifi will generate objects with well-formed UUIDv4 guids and
the webhook will replace thoes of them that need to be prefixed. The
webhook can be switched on and off.

Note: We had to generate new guids for some sercrets/deestinations,
where previously we were reusing the parent object guid, but all the e2e
tests passed with the change.

Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>
Co-authored-by: Danail Branekov <danailster@gmail.com>