Create a uniques app whenenver we bind/unbind to a service instance
Before this change all tests that were binding to as service instances
were using the buildpack app. When we bind to an app that is ready and
running the app becomes unready for a while and restarts. If we try to
run a task against such an app the operation will fail, because the app
is unready. This change creates a cheap app record and binds/unbinds to
this app to prevent messing with the running app state

Co-authored-by: Yusmen Zabanov <yusmen.zabanov@sap.com>