Add process guid to gauge envelopes
The log cache api will be used by the process and app handlers, so that they
can fetch stats for a process with a particular guid. Therefore it makes
sense to filter the envelope stream for envelopes for that process guid
only. In order to do so we need to have access to the process guid that
the envelope is related to.

The cloud controller implementation is doing the same: https://github.com/cloudfoundry/cloud_controller_ng/blob/5b1af263bbb21793ba7d49f64c75f6ab5f5ab267/lib/cloud_controller/diego/reporters/instances_stats_reporter.rb#L144

Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>