Fix a binding controller flake
Flake: https://ci.korifi.cf-app.com/teams/main/pipelines/main/jobs/run-tests-periodic/builds/22068

According to failure logs, the service binding is being reconciled
multiple times before the binding gets into its ready state.

This means that if we stub the broker client to return a sync response
on invocation 3 only, it is quite probable that the fake's sync response
is not properly processed and subsequent reconciles still see that the
operation is ongoing.

Therefore, this change alters stubs the broker client to return a sync
response (i.e. override the default async one) when it returns a
succeeded operation