WIP
Attempt to address the flake from the previous commit

* When creating/deleting a binding, the binding repository awaits not
  only the app ready state, but also checks whether the binding guid
  - has been added to both the spec and the status (on create)
  - has been removed from both spec and the status (on delete)

We do this in order to ensure that the controller has "seen" the change
the repository is trying to perform. Only waiting for the ready state
seems to not be reliable enough as the app controller might have
reconciled an outdated state, thus making the app ready, but not
reflecting the change the repository is trying to perform.

TODO: the env bit from the apps e2e tests fail - probably a bug in the
spike implementation

TODO: check whether this solves the flake. It is easy to reproduce if
e2e apps tests are focused and run in parallel in two shells.

Co-authored-by: Yusmen Zabanov <yusmen.zabanov@sap.com>