Set `AppWorkload.Status.Instances` to statefulset's `readyReplicas`
According to the `StatefulSet.Status` documentation, `Replicas` is the
number of pods the statefulset controller has created, while
`ReadyReplicas` is the number of statefulset pods that have their
`Ready` condition set.

Previously, when we used to set `AppWorkload.Status.Instances`, the app
could become `ready` shortly before the pods are working and the app can
be requested. By using `ReadyReplicas` instead, we guarantee that
whenever the app is reported as `started`, it can be immediately
requested.

issue #3825

Co-authored-by: Danail Branekov <danailster@gmail.com>