Ignore NotFound error on deleting service instances
When deleting service instances, the service instance repository would
first get the service instance, if purging is requested - remove
finalizers, and then delete the instance.

At the point in time when deletion from k8s is performed, failing the
delete operation on not-found error makes no sense - the instance is
gone anyway.

In case the user tries to delete an instance that never existed, getting
the service instance would fail with not-found, and the user would get
the error.

Co-authored-by: Georgi Sabev <georgethebeatle@gmail.com>