Controllers do not send empty patches to k8s api server
Korifi controllers are based on the "patching reconciler" - at the end
of each `ReconcileResource` method, they patch the resource with the api
server. This results into lots of unnecessary empty patches being sent.

In order to address this,
* Introduce a `client.Client` wrapper that ignores empty patches on
  `Patch` and `Status().Patch`
* If the patch is not empty, the request is delegated to the original
  k8s client.
* All other methods of the `client.Client` interface are plain
  delegation

This new wrapper is used when wiring controllers with k8s client in the
respective mains

fixes #3807

Co-authored-by: Yusmen Zabanov <yusmen.zabanov@sap.com>