WIP
TODO
In the previous commit we used the space-guid webhook to also set guids.
It seems that this is not a correct decision, as only space-scoped
objects have space guids while ALL CF objects (both space-scoped and
deployment-scoped, such as orgs) have guids.
We should therefore probably have two webhooks - one that sets the
space-guids, and another one that sets the guids. Initially we could
have the latter to only set the cf app guid. We could add more resources
as we work on the story

Co-authored-by: Danail Branekov <danailster@gmail.com>