WIP: Works on kind too
The yq caBundle injection in the Makefile has become a monster though.
We have to do the secret lookup only when generation is disabled. Lookup
requires that the secrets are available before the korifi installation.
That's why in the cert manager case we have to fall back to "Cg==",
which is the base64 version of `""`. In the cert manager case the ca is
dynamically injected by cert manager.

This one-liner could potentially be extracted into a template. This will
make it simpler to call, while in the template we could expand it to
more than one line for readability

Alternatively we could derop lookup and go back to passing the PEM value
to the helm chart (not the secret name). Passing the secret name is a
better experience to the chart user though. If we go back to PEM this
would mean that whoever is installing korifi without cert manager will
have to pull the ca out of the secret themselves

Co-authored-by: Yusmen Zabanov <yusmen.zabanov@sap.com>